refacto org settings page
