+++ app/lib/ai-providers.ts
@@ -3,9 +3,13 @@
  * Ce fichier peut être importé côté client et serveur.
  */
 
-export type AiProvider = "openai" | "gemini";
+export enum AiProviderEnum {
+  OPENAI = "openai",
+  GEMINI = "gemini",
+}
 
-export const AI_PROVIDERS: { value: AiProvider; label: string }[] = [
-  { value: "openai", label: "OpenAI (GPT)" },
-  { value: "gemini", label: "Google Gemini" },
+// TODO use a ICU translated value and use only the enum
+export const AI_PROVIDERS: Array<{ value: AiProviderEnum; label: string }> = [
+  { value: AiProviderEnum.OPENAI, label: "OpenAI (GPT)" },
+  { value: AiProviderEnum.GEMINI, label: "Google Gemini" },
 ];
