+++ app/routes/orgs.$orgSlug.projects.$projectSlug.keys.$keyId.tsx
@@ -57,6 +57,7 @@ import {
 import { toaster } from "~/components/ui/toaster";
 import type { TranslationSuggestion } from "~/lib/ai-translation.server";
 import { getInstance } from "~/middleware/i18next";
+import { AiProviderEnum } from "~/lib/ai-providers";
 
 export async function loader({ request, params }: Route.LoaderArgs) {
   const user = await requireUser(request);
@@ -584,8 +585,9 @@ export default function EditTranslationKey({
                       {aiFetcher.data.provider && (
                         <Text fontSize="xs" color="gray.500">
                           Généré par{" "}
-                          {aiFetcher.data.provider === "openai"
-                            ? "OpenAI"
+                          {aiFetcher.data.provider === AiProviderEnum.OPENAI
+                            ? // TODO use translation switch like in ~lib/ai-providers.ts
+                              "OpenAI"
                             : "Google Gemini"}
                         </Text>
                       )}
